feat(ui): a package the primitives can be drawn from twice (H21)
The control plane writes its own dialog, button and input because ours could
not be imported. `packages/web` is a private application with no `exports`, so
nothing outside this repo resolves a subpath from it, and a consumer taking it
into a workspace would inherit `@hezo/server` along with Radix, the router,
recharts and react-markdown — none of which a control plane has a use for.

Three things made the primitives unshareable, and none was large.

**A primitive resolved its own copy.** `DialogContent` looked up `common.close`
and `ConfirmDialog` `common.cancel`, and `useI18n` throws rather than falling
back — so rendering either outside an app holding a catalog was impossible, for
two words. Every user-visible string is a prop with an English default now, and
`components/ui/` keeps the lookups.

**A class string read a raw custom property.** `bg-[var(--overlay)]` is defined
only where a stylesheet happens not to namespace its raw tokens; where one does,
the backdrop is transparent and the markup says nothing. `bg-overlay` reads the
key `@theme` publishes, which every consumer defines.

**The stacking was this app's.** The z-indices were chosen against the
task-detail preview panel and the review bottom sheet. They are named exports a
consumer overrides through the `className` already appended last.

Nothing moved for a caller: `components/ui/*` are pass-throughs and the two
dialogs are wrappers, so all 1852 web specs pass untouched and no catalog
changed. `test/ui-package-standalone.test.tsx` renders the primitives with no
`I18nProvider` — the one thing every other spec in the tree, which always has
one, cannot catch. Both regressions were confirmed by reverting each fix and
watching it go red.

Docs-Checked: added `packages/ui` to the layout and its rules to
  `.dev/architecture.md`, a row to `.dev/seam-registry.md` and the seam excerpt
  in `AGENTS.md`, and recorded H21 done in `.dev/hezo-cloud-requirements.md`
Translations-Checked: no string changed — the primitives take labels as props
  and the app wrappers still pass `common.close` and `common.cancel`, so all
  twelve catalogs are untouched
