+++ .dev/seam-registry.md
@@ -48,6 +48,7 @@ rather than here, is how a codebase ends up with two of everything.
 | Open/close motion for a right-hand side panel | `--panel-motion` + `.panel-enter` / `.panel-exit` (`packages/web/src/index.css`), reached through `packages/web/src/lib/panel-motion.ts`. A panel that mounts when it opens takes the animations; one that is always mounted and only slides takes `PANEL_MOTION_TRANSITION`. `ResizableSplit` arms its grid-track transition on the panel appearing or disappearing and disarms it after - never leave it on, or every divider drag and Arrow-key step lags by the full beat |
 | A breadcrumb row - one line, scrolling sideways rather than wrapping or truncating | `BreadcrumbRow` (`components/ui/breadcrumb.tsx`) - `Breadcrumb` renders through it; a caller with its own links takes the row, never a second `<nav>` |
 | A device sign-in (copy a one-time code, enter it on a provider's page) | `DeviceCodeSteps` (`components/ui/device-code-steps.tsx`) - callers supply transport only |
+| A primitive more than one app draws with (dialog, confirmation, button, input, keycap, shortcut binding) | `@hezo/ui` (`packages/ui/src/`) - the app's `components/ui/` files are pass-throughs, and the two dialogs are wrappers supplying the translated labels. Never resolve copy or read a raw `var(--token)` inside the package: either makes it unimportable, silently |
 | An optimistic mutation | `useOptimisticMutation` (`hooks/use-optimistic-mutation.ts`) |
 | A server test context | `createTestContext()` (`test/helpers/context.ts`) |
 | A migration test | `createDataPreservationHarness()` (`test/helpers/migrate.ts`) |
