+++ packages/ui/package.json
@@ -1,0 +1,24 @@
+{
+	"name": "@hezo/ui",
+	"version": "0.57.0",
+	"type": "module",
+	"exports": {
+		".": "./src/index.ts"
+	},
+	"main": "./src/index.ts",
+	"scripts": {
+		"typecheck": "tsc -p tsconfig.json"
+	},
+	"dependencies": {
+		"@radix-ui/react-alert-dialog": "^1",
+		"@radix-ui/react-dialog": "^1",
+		"lucide-react": "^0.500"
+	},
+	"peerDependencies": {
+		"react": "^19"
+	},
+	"devDependencies": {
+		"@types/react": "^19",
+		"typescript": "^5"
+	}
+}
